Remove FinremCaseDataParent
